#15 - Move tool runners to their own sub package + fix mock file names
