+++ .github/workflows/CI.yml
@@ -70,7 +70,7 @@ jobs:
         go-version: 1.16
     
     - name: Run Unit Tests
-      run: go test -v -race -covermode=atomic -coverprofile=coverage.out ./controllers/ ./controllers/password_helpers/ ./data/ ./models/ ./router/ ./router/handlers/ ./server/ ./tools/admin_creator/ ./tools/migration_runner/
+      run: go test -v -race -covermode=atomic -coverprofile=coverage.out ./controllers ./controllers/password_helpers ./data ./models ./router ./router/handlers ./server ./tools/admin_creator/runner ./tools/migration_runner/runner
 
     - name: Install Goveralls
       run: go install github.com/mattn/goveralls@latest
