+++ tools/admin_creator/main.go
@@ -1,13 +1,10 @@
 package main
 
 import (
-	"authserver/common"
 	"authserver/config"
-	"authserver/controllers"
-	"authserver/data"
 	"authserver/dependencies"
+	"authserver/tools/admin_creator/runner"
 	"flag"
-	"fmt"
 	"log"
 
 	"github.com/spf13/viper"
@@ -27,24 +24,8 @@ func main() {
 
 	viper.Set("db_key", *dbKey)
 
-	err = Run(dependencies.ResolveScopeFactory(), dependencies.ResolveControllers(), *username, *password)
+	err = runner.Run(dependencies.ResolveScopeFactory(), dependencies.ResolveControllers(), *username, *password)
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-// Run runs the admin creator. Returns any errors
-func Run(sf data.ScopeFactory, c controllers.UserController, username string, password string) error {
-	return sf.CreateDataExecutorScope(func(exec data.DataExecutor) error {
-		return sf.CreateTransactionScope(exec, func(tx data.Transaction) (bool, error) {
-			//save the user
-			user, rerr := c.CreateUser(tx, username, password)
-			if rerr.Type != common.ErrorTypeNone {
-				return false, common.ChainError("error creating user", rerr)
-			}
-
-			fmt.Println("Created user:", user.ID.String())
-			return true, nil
-		})
-	})
 }
