+++ tools/migration_runner/main.go
@@ -1,11 +1,9 @@
 package main
 
 import (
-	"authserver/common"
 	"authserver/config"
-	"authserver/data"
 	"authserver/dependencies"
-	"authserver/tools/migration_runner/interfaces"
+	"authserver/tools/migration_runner/runner"
 	"flag"
 	"log"
 
@@ -25,32 +23,12 @@ func main() {
 
 	viper.Set("db_key", *dbKey)
 
-	mrf := interfaces.CoreMigrationRunnerFactory{
+	mrf := runner.CoreMigrationRunnerFactory{
 		MigrationRepositoryFactory: dependencies.ResolveMigrationRepositoryFactory(),
 	}
 
-	err = Run(dependencies.ResolveScopeFactory(), &mrf, *down)
+	err = runner.Run(dependencies.ResolveScopeFactory(), &mrf, *down)
 	if err != nil {
 		log.Fatal(err)
 	}
-}
-
-// Run runs the migration runner. Returns any errors.
-func Run(sf data.ScopeFactory, mrf interfaces.MigrationRunnerFactory, down bool) error {
-	return sf.CreateDataExecutorScope(func(exec data.DataExecutor) error {
-		mr := mrf.CreateMigrationRunner(exec)
-		var err error
-
-		//run the migrations
-		if down {
-			err = mr.MigrateDown()
-		} else {
-			err = mr.MigrateUp()
-		}
-
-		if err != nil {
-			return common.ChainError("error running migrations", err)
-		}
-		return nil
-	})
 }
